+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Add button to show active workspace layout and click to cycle through available layouts.
 
+### Changed
+
+- Active workspace indicator will no longer appear on top of the switcher button and will always be at the bottom of it.
+
 ## [0.9.2] - 2026-01-06
 
 ### Fixed